+++ app/controllers/concerns/auth.rb
@@ -1,5 +1,3 @@
-class UserNotAuthorised < StandardError; end
-
 module Auth
   extend ActiveSupport::Concern
 
@@ -7,13 +5,8 @@ module Auth
     helper_method :current_user
     helper_method :authenticated?
 
-    rescue_from(UserNotAuthorised, Pundit::NotAuthorizedError) do |exception|
-      error_message = if exception.respond_to?(:policy)
-        t("#{exception.policy.class.to_s.underscore}.#{exception.query}", scope: "not_authorised", default: :default)
-      else
-        t("page_content.errors.not_authorised.explanation")
-      end
-
+    rescue_from(Pundit::NotAuthorizedError) do |exception|
+      error_message = t("page_content.errors.not_authorised.explanation")
       render "pages/errors/not_authorised", formats: [:html], status: 401, locals: {error_message: error_message}
     end
   end