+++ spec/models/activity_spec.rb
@@ -135,17 +135,20 @@ RSpec.describe Activity, type: :model do
     describe ".comments" do
       it "fetches comments related to a given activity" do
         activity = create(:project_activity)
+        actual = create(:actual, :with_comment, parent_activity: activity)
         refund = create(:refund, parent_activity: activity)
         adjustment = create(:adjustment, parent_activity: activity)
 
         activity_comment = create(:comment, commentable: activity)
 
         _comments = create_list(:comment, 5)
+        _actuals = create_list(:actual, 2, :with_comment)
         _refunds = create_list(:refund, 2)
         _adjustments = create_list(:adjustment, 2)
 
         expect(activity.comments).to match_array([
           activity_comment,
+          actual.comment,
           refund.comment,
           adjustment.comment
         ])